ft: code changes to handle processor concurrency
- put queue processor task state in a QueueProcessorTask class
  instance, separate from the main QueueProcessor class which is now
  able to process multiple entries at once

- rework _getLocations helper: merge it into QueueEntry class as
  getReducedLocations() to ease usage and testing. Unit tests are also
  merged into QueueEntry.spec.js

- make the concurrency configurable, default is 10

- create a cache for vault clients to re-use them across requests and
  benefit from keep-alive

- use a cache for vault clients so that they can be re-used across
  processing tasks
