+++ conf/config.joi.js
@@ -55,6 +55,8 @@ const joiSchema = {
             queueProcessor: {
                 groupId: joi.string().required(),
                 retryTimeoutS: joi.number().default(300),
+                // versioning can support out of order updates
+                concurrency: joi.number().greater(0).default(10),
             },
         },
     },
